Refactor metadata viewing and related functions to reduce duplication
